Database update needed:  Implemented ship skinning services. These may still need some work but should be adequate for now.
The skinShip table provided by the fuzzworks database export has an error that needs to be fixed.
Newer versions of the database do not have this error so this fix will not be needed in the future.
To fix an existing database use the command:
  source evemu_server/sql/ofic-updates/Fix-skinShip.sql
This will be automatic on full reinstall of database server as the merge-ofic-updates.(sh|bat) command will include this file automatically.

Two new tables have been added.  These are included in the evemu_dynamic-dump.sql file.
To add the new tables without resetting the database use the following commands:
DROP TABLE IF EXISTS `srvChrSkins`;
CREATE TABLE `srvChrSkins`(
  `characterID` INT(10) UNSIGNED NOT NULL DEFAULT '0',
  `skinID` INT(10) UNSIGNED NOT NULL default '0',
  `expireTime` BIGINT(20) UNSIGNED DEFAULT NULL,
  PRIMARY KEY (`characterID`, `skinID`)
);
DROP TABLE IF EXISTS `srvShipSkin`;
CREATE TABLE `srvShipSkin` (
  `itemID` INT(10) UNSIGNED NOT NULL default '0',
  `skinID` INT(10) UNSIGNED NOT NULL default '0',
  PRIMARY KEY (`itemID`)
);